+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [To be deprecated]
-- Remove "/features" api path.
+- Remove deprecated "/mocks" api path.
 
 ## [unreleased]
 ### Added
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 
 ## [1.2.0] - 2019-12-15
+### Added
+- Add "adminApiPath" option, which allows to change the new API path ("/admin").
+- Add "adminApiDeprecatedPaths" option, which allows to disable the deprecated API path ("/mocks").
 
 ## [1.1.0] - 2019-12-07
 ### Changed